my_env/requirements.txt
openenv-core>=0.2.3
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.0.0